requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
sqlalchemy-libsql==0.2.0
python-dotenv==1.0.1
